--- src/pubsub/Event.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <string>
-
-class Event
-{
-public:
-	virtual ~Event() {};
-};
