+++ src/pubsub/PubSub.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-#include "Event.hpp"
 #include "EventManager.hpp"
+#include "ECSEvents.hpp"
